Move problems over to component, and make it only cover the update/create section
